Move resources and pages to resources dir
