+++ src/Resources/RoleResource/Pages/CreateRole.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Filament\Resources\Shield\RoleResource\Pages;
+namespace BezhanSalleh\FilamentShield\Resources\RoleResource\Pages;
 
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 use Spatie\Permission\Models\Permission;
 use Filament\Resources\Pages\CreateRecord;
-use App\Filament\Resources\Shield\RoleResource;
+use BezhanSalleh\FilamentShield\Resources\RoleResource;
 use Illuminate\Support\Collection;
 
 class CreateRole extends CreateRecord
